Section 3 - Camera & Aim
Updated some visuals for the camera to follow the player around the scene, as well as a look-ahead feature based on the player's movement. Created a placeholder bullet and laser for future development of the weapon system. Implemented target lock and precise aim features for more immersion.  Due to a recommendation from Unity, the editor version was updated on this commit from 2022.3.2 to 2022.3.62.